# Optimize DOM HTML serialization for UTF-8